refactor pagination module
