+++ ninja_extra/pagination/models/__init__.py
@@ -1,0 +1,5 @@
+from .page_by_number import PageNumberPaginationExtra
+
+__all__ = [
+    "PageNumberPaginationExtra",
+]
